# Backend/jiufang-project/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: 九方智能查询系统 API
  description: |
    九方智能查询系统是一个基于AI Agent的自然语言查询系统，支持用户通过自然语言查询ERP数据，
    并提供定时报告、预警规则、数据导出等功能。
    
    ## 认证方式
    - 使用JWT Bearer Token认证
    - 登录接口获取JWT令牌
    - 令牌有效期24小时
    
    ## 权限控制
    - admin: 系统管理员，拥有所有权限
    - manager: 管理层，拥有查询和导出权限
    - executive: 执行层，拥有查询权限
    
    ## 响应格式
    所有接口统一使用以下响应格式：
    ```json
    {
      "code": 200,
      "message": "success",
      "data": {}
    }
    ```
  version: 1.0.0
  contact:
    name: 九方技术团队
    email: tech@jiufang.com
servers:
  - url: https://api.jiufang.com/v1
    description: 生产环境
  - url: https://api-dev.jiufang.com/v1
    description: 开发环境
  - url: http://localhost:8080/api/v1
    description: 本地开发环境
...
paths:
...
  /dialog-favorites:
    post:
      tags:
        - DialogFavorite
      summary: 创建对话收藏
      description: 创建对话收藏
      operationId: createDialogFavorite
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - session_id
              properties:
                session_id:
                  type: string
                  pattern: '^\d{19}$'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DialogFavoriteResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    get:
      tags:
        - DialogFavorite
      summary: 查询对话收藏列表
      description: 分页查询用户的对话收藏列表
      operationId: getDialogFavoriteList
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DialogFavoriteListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /dialog-favorites/{favorite_id}:
    delete:
      tags:
        - DialogFavorite
      summary: 删除对话收藏
      description: 删除对话收藏
      operationId: deleteDialogFavorite
      security:
        - bearerAuth: []
      parameters:
        - name: favorite_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^\d{19}$'
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== 导出模块 ====================
...
  /alerts:
    post:
      tags:
        - Alert
      summary: 创建预警规则
      description: 创建预警规则（仅系统管理员可操作）
      operationId: createAlert
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - condition
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                condition:
                  type: string
                  description: 预警条件SQL表达式
                threshold:
                  type: number
                recipients:
                  type: array
                  items:
                    type: string
                    format: email
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    get:
      tags:
        - Alert
      summary: 查询预警规则列表
      description: 分页查询预警规则列表（仅系统管理员可操作）
      operationId: listAlerts
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /alerts/{id}:
    get:
      tags:
        - Alert
      summary: 查询预警规则详情
      description: 查询预警规则详情（仅系统管理员可操作）
      operationId: getAlert
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^\d{19}$'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Alert
      summary: 更新预警规则
      description: 更新预警规则（仅系统管理员可操作）
      operationId: updateAlert
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^\d{19}$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                condition:
                  type: string
                threshold:
                  type: number
                recipients:
                  type: array
                  items:
                    type: string
                    format: email
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Alert
      summary: 删除预警规则
      description: 删除预警规则（仅系统管理员可操作）
      operationId: deleteAlert
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^\d{19}$'
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== 反馈模块 ====================
...
components:
  schemas:
    SuccessResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: 错误码
          example: 40001
        message:
          type: string
          description: 错误信息
          example: "参数错误"
...
    DialogFavoriteResponse:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          required:
            - id
            - session_id
            - created_at
          properties:
            id:
              type: string
              pattern: '^\d{19}$'
            session_id:
              type: string
              pattern: '^\d{19}$'
            created_at:
              type: string
              format: date-time
    DialogFavoriteListResponse:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          required:
            - total
            - page
            - page_size
            - favorites
          properties:
            total:
              type: integer
            page:
              type: integer
            page_size:
              type: integer
            favorites:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                    pattern: '^\d{19}$'
                  session_id:
                    type: string
                    pattern: '^\d{19}$'
                  created_at:
                    type: string
                    format: date-time
...
    AlertResponse:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          required:
            - id
            - name
            - condition
            - created_at
          properties:
            id:
              type: string
              pattern: '^\d{19}$'
            name:
              type: string
            condition:
              type: string
            threshold:
              type: number
            recipients:
              type: array
              items:
                type: string
                format: email
            created_at:
              type: string
              format: date-time
    AlertListResponse:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          required:
            - total
            - page
            - page_size
            - alerts
          properties:
            total:
              type: integer
            page:
              type: integer
            page_size:
              type: integer
            alerts:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                    pattern: '^\d{19}$'
                  name:
                    type: string
                  condition:
                    type: string
                  created_at:
                    type: string
                    format: date-time
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 40001
            message: "参数错误"
    Unauthorized:
      description: 未授权
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 40101
            message: "未授权访问"
    Forbidden:
      description: 权限不足
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 40301
            message: "权限不足"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 40401
            message: "资源不存在"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT认证